Struct icaros_base::comms::air::scheduler::AirPacketScheduler

pub struct AirPacketScheduler<T, S, Tx, Rx> where
    T: AirCommunicationService<Tx, Rx>,  {
    comm_service: T,
    tx_queue: VecDeque<Tx>,
    tx_app_queue: VecDeque<ApplicationPacket>,
    rx_app_queue: VecDeque<ApplicationPacket>,
    latency_recv: u32,
    pkt_counter: u16,
    last_gps_data_send: Instant,
    _marker_state: PhantomData<S>,
    _marker_rx: PhantomData<Rx>,
}

Packet scheduler for ground-to-air and air-to-ground communications

T defines the underlying device driver to which packets are scheduled

S defines the storage type to where the scheduler will go to fetch/store data

Tx defines the type of the messages being sent through the device driver

Rx defines the type of the messages being received through the device driver

This packet scheduler holds several implementations, one for each side of the communication.

Fields
comm_service: Ttx_queue: VecDeque<Tx>tx_app_queue: VecDeque<ApplicationPacket>rx_app_queue: VecDeque<ApplicationPacket>latency_recv: u32pkt_counter: u16last_gps_data_send: Instant_marker_state: PhantomData<S>_marker_rx: PhantomData<Rx>
Methods

impl<T, S, Tx, Rx> AirPacketScheduler<T, S, Tx, Rx> where
T: AirCommunicationService<Tx, Rx>,

Common routines for all implementations

fn increment_pkt_counter(&mut self)

Increments the packet counter for the next packet to be sent

impl<T, S> AirPacketScheduler<T, S, A2GMessage, G2AMessage> where
T: AirCommunicationService<A2GMessage, G2AMessage>,
S: AirState,

Implementation-specific methods for implementation of the packet scheduler for the drone side and its corresponding device driver and state storage implementations.

pub fn initialize(
config: T::AirCommunicationOptions
) -> Result<Self, T::HardwareDriverError>

Initializes a new air-side packet scheduler, together with its underlying driver and given the latter's settings to be used

fn generate_message_from_state(&mut self, state: &S) where
S: AirState,

Generates a set of messages that convey the given state and schedules them for sending

fn update_state_from_message(&mut self, state: &mut S, message: G2AMessage) where
S: AirState,

Updates the given system state from the given incoming message

impl<T, S> AirPacketScheduler<T, S, G2AMessage, A2GMessage> where
T: AirCommunicationService<G2AMessage, A2GMessage>,
S: GroundState,

Implementation-specific methods for implementation of the packet scheduler for the ground control side and its corresponding device driver and state storage implementations.

pub fn initialize(
config: T::AirCommunicationOptions
) -> Result<Self, T::HardwareDriverError>

Initializes a new ground-side packet scheduler, together with its underlying driver and given the latter's settings to be used

fn generate_message_from_state(&mut self, state: &S) where
S: GroundState,

fn update_state_from_message(&mut self, state: &mut S, message: A2GMessage) where
S: GroundState,