Trait icaros_base::comms::air::scheduler::AirState

pub trait AirState {
    fn get_motor_speeds(&self) -> MotorSpeed;

    fn get_orientation(&self) -> Orientation;

    fn get_acceleration(&self) -> Acceleration;

    fn set_controller_state(&mut self, new: G2AControllerAxisState);

    fn get_gps_coordinates(&self) -> Option<FullGPSData>;

    fn set_gps_coordinates(
        &mut self, 
        new_latlon: (Option<f64>, Option<f64>), 
        alt: Option<f32>
    );
}

A storage for drone in-flight state

Implementors are able to function as storage for the drone's in-flight state where the packet schedulers can store data on. This could be a memory-only structure or a simple database with logging capabilities, depending on the use case.

Required methods

fn get_motor_speeds(&self) -> MotorSpeed

Returns the drone's current motor speeds

fn get_orientation(&self) -> Orientation

Returns the drone's current absolute orientation

fn get_acceleration(&self) -> Acceleration

Returns the drone's current linear acceleration

fn set_controller_state(&mut self, new: G2AControllerAxisState)

Stores the current gamepad state received from ground control

fn get_gps_coordinates(&self) -> Option<FullGPSData>

If full GPS data is available, returns it; otherwise returns None

fn set_gps_coordinates(
&mut self,
new_latlon: (Option<f64>, Option<f64>),
alt: Option<f32>
)

Stores the provided GPS data

Loading content...