Trait icaros_base::comms::air::AirCommunicationService

pub trait AirCommunicationService<Tx, Rx>: Sized {
    type AirCommunicationOptions;
    type HardwareDriverError: 'static + Error;
    fn setup(
        config: Self::AirCommunicationOptions
    ) -> Result<Self, Self::HardwareDriverError>;

    fn get_max_app_message_size() -> usize;

    fn is_tx_busy(&mut self) -> bool;

    fn send(
        &mut self, 
        msg: Tx
    ) -> Result<bool, SendError<Self::HardwareDriverError>>;

    fn recv_available(&mut self) -> usize;

    fn recv(&mut self) -> Result<Rx, ReceiveError<Self::HardwareDriverError>>;
}

An air-to-ground communication driver

This trait, when implemented, allows the implementor to act as a device driver for communications with the ground control.

Tx is the type of the messages being transmitted Rx is the type of the messages being received

Associated Types

type AirCommunicationOptions

The settings structure that has to be passed to the device driver

type HardwareDriverError: 'static + Error

The error type definining hardware-specific errors

Loading content...
Required methods

fn setup(
config: Self::AirCommunicationOptions
) -> Result<Self, Self::HardwareDriverError>

Creates a new instance of the implementor with the given configuration

fn get_max_app_message_size() -> usize

Returns the maximum allowed size of application packets passing through this driver

fn is_tx_busy(&mut self) -> bool

Returns whether or not the underlying device is not yet ready for sending packets

fn send(
&mut self,
msg: Tx
) -> Result<bool, SendError<Self::HardwareDriverError>>

Sends a single message through the device

fn recv_available(&mut self) -> usize

Returns how many (or, depending on implementation, whether any) messages are available

fn recv(&mut self) -> Result<Rx, ReceiveError<Self::HardwareDriverError>>

Receives a single message from the device

Loading content...