Trait icaros_base::comms::common::ApplicationPacketScheduler

pub trait ApplicationPacketScheduler {
    fn send_app_packet(
        &mut self, 
        data: ApplicationPacket
    ) -> Result<bool, Box<dyn Error>>;

    fn recv_app_packets(
        &mut self
    ) -> Result<&mut VecDeque<ApplicationPacket>, Box<dyn Error>>;
}

A packet scheduler for application packets.

Implementors of this trait can act as application packet schedulers, middleware between device drivers and the core systems whose purpose is to schedule the input/output of application packets.

Required methods

fn send_app_packet(
&mut self,
data: ApplicationPacket
) -> Result<bool, Box<dyn Error>>

Schedule for and send the given application packet to the device driver

fn recv_app_packets(
&mut self
) -> Result<&mut VecDeque<ApplicationPacket>, Box<dyn Error>>

Receive application packets from the device driver

Loading content...