Trait icaros_base::comms::common::PacketScheduler

pub trait PacketScheduler<I, O, S> {
    fn can_send(&mut self) -> bool;

    fn send_state(&mut self, state: &S) -> Result<bool, Box<dyn Error>>;

    fn recv_state(&mut self, state: &mut S) -> Result<bool, Box<dyn Error>>;
}

A packet scheduler.

Implementors of this trait can act as packet schedulers, middleware between device drivers and the core systems whose purpose is to schedule the input/output of packets.

I represents the type of the incoming packet (i.e. packet that updates state) O represents the type of the outgoing packet (i.e. packet that replicates state) S represents the type of the state being written or read to/from.

Required methods

fn can_send(&mut self) -> bool

Returns whether packets can be sent through this packet scheduler

fn send_state(&mut self, state: &S) -> Result<bool, Box<dyn Error>>

Schedule for and send the given state to the device driver

fn recv_state(&mut self, state: &mut S) -> Result<bool, Box<dyn Error>>

Receive data from the device driver and update the current state

Loading content...