Trait icaros_base::comms::controller::ControllerCommunicationService

pub trait ControllerCommunicationService: Sized {
    type ControllerCommunicationOptions;
    type HardwareDriverError: 'static + Error;
    fn setup(
        config: Self::ControllerCommunicationOptions
    ) -> Result<Self, Self::HardwareDriverError>;

    fn is_tx_busy(&mut self) -> bool;

    fn send(
        &mut self, 
        msg: R2AMessage
    ) -> Result<bool, ControllerSendError<Self::HardwareDriverError>>;

    fn recv_available(&mut self) -> Result<usize, Self::HardwareDriverError>;

    fn recv(
        &mut self
    ) -> Result<A2RMessage, ControllerReceiveError<Self::HardwareDriverError>>;
}

A flight controller communication driver

This trait, when implemented, allows the implementor to act as a device driver for communications with the flight controller.

Associated Types

type ControllerCommunicationOptions

The settings structure that has to be passed to the device driver

type HardwareDriverError: 'static + Error

The error type definining hardware-specific errors

Loading content...
Required methods

fn setup(
config: Self::ControllerCommunicationOptions
) -> Result<Self, Self::HardwareDriverError>

Creates a new instance of the implementor with the given configuration

fn is_tx_busy(&mut self) -> bool

Returns whether or not the underlying device is not yet ready for sending packets

fn send(
&mut self,
msg: R2AMessage
) -> Result<bool, ControllerSendError<Self::HardwareDriverError>>

Sends a single message through the device

fn recv_available(&mut self) -> Result<usize, Self::HardwareDriverError>

Returns how many (or, depending on implementation, whether any) messages are available

fn recv(
&mut self
) -> Result<A2RMessage, ControllerReceiveError<Self::HardwareDriverError>>

Receives a single message from the device

Loading content...