Module icaros_base::utils

Utility structures and functions

Re-exports

pub use quartenion::Quartenion;

pub use vector::Point3;

Modules

crc16
CRC calculation utilities

quartenion
A quartenion, used to represent 3D rotation.

vector
A 3-dimensional vector

Functions

is_all_same
Returns true if all the elements on a given array are the same

map_i8_into_u8
Maps a signed 1-byte value into an unsigned 1-byte value

map_values
Maps a given value X from a given input range to the given output range

read_le_u16
Reads an unsigned 2-byte number out of a byte array