Module icaros_base::utils::crc16

CRC calculation utilities

Source: From 1 which by itself cites:

See the Dallas Semiconductor app note 27 for 8051 assembler example and general CRC optimization suggestions. The table on the last page of the app note is the key to understanding these implementations.

Jack Crenshaw's "Implementing CRCs" article in the January 1992 isue of Embedded Systems Programming. This may be difficult to find, but it explains CRC's in very clear and concise terms. Well worth the effort to obtain a copy.

https://www.nongnu.org/avr-libc/user-manual/group__util__crc.html ↩

Functions

_crc16_update
CRC calculation update function mimicking the one built into AVR processors' standard libraries (i.e. Arduino)

crc16_calculate
Calculates the CRC checksum of a given array